refactor(transfer): split generator file_list entry into submodules
