+++ crates/transfer/src/generator/file_list/entry/device.rs
@@ -1,0 +1,31 @@
+//! Device number extraction for `FileEntry` rdev fields.
+//!
+//! Splits a raw `rdev` value into major/minor components using the
+//! platform-specific bit layout (Linux split encoding vs BSD/macOS).
+
+/// Extracts major and minor device numbers from a raw `rdev` value.
+///
+/// The layout differs by platform:
+/// - **Linux**: Split encoding where major/minor span non-contiguous bits.
+/// - **macOS/BSD**: Major in high byte, minor in low 24 bits.
+///
+/// # Upstream Reference
+///
+/// Mirrors glibc `major()`/`minor()` macros used by upstream rsync to populate
+/// `rdev_major`/`rdev_minor` in `file_struct`.
+#[cfg(all(unix, target_os = "linux"))]
+pub(in crate::generator) fn rdev_to_major_minor(rdev: u64) -> (u32, u32) {
+    let major = ((rdev >> 8) & 0xfff) as u32 | (((rdev >> 32) & !0xfff) as u32);
+    let minor = (rdev & 0xff) as u32 | (((rdev >> 12) & !0xff) as u32);
+    (major, minor)
+}
+
+/// Extracts major and minor device numbers from a raw `rdev` value (BSD/macOS).
+///
+/// BSD layout: major in bits 31-24, minor in bits 23-0.
+#[cfg(all(unix, not(target_os = "linux")))]
+pub(in crate::generator) fn rdev_to_major_minor(rdev: u64) -> (u32, u32) {
+    let major = (rdev >> 24) as u32;
+    let minor = (rdev & 0xffffff) as u32;
+    (major, minor)
+}
